Avances métodos

+++ database/migrations/2023_06_07_194209_create_pokemons_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('pokemons', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('height');
-            $table->integer('weight');
+            $table->decimal('height');
+            $table->decimal('weight');
             $table->timestamps();
         });
     }
